Fix #1347, dispatch if last masks xts::last
data.table::last is intended to work with xts::last, which is a S3
generic. e011351 appears to have caused a regression that prevents
xts::last from dispatching.

Revert e011351, but keep the is.data.frame check on the happy path, so
it will be fast for data.table objects. Return xts::last if xts is on
the search path. Otherwise, return tail(x, ...) instead of throwing an
error (the previous behavior).